fixed bug with dissapearing labels, added separate frame class
