+++ lab4/resources/config.properties
@@ -4,7 +4,7 @@ body_storage_size=100
 engine_storage_size=100
 workers_amount=1
 dealers_amount=1
-car_accessory_suppliers_amount=1
+car_accessory_suppliers_amount=2
 car_body_suppliers_amount=1
 car_engine_suppliers_amount=1
